Extract StayInformedRedesignedComponent
