+++ ddp-workspace/projects/toolkit/src/public-api.ts
@@ -36,7 +36,8 @@ export * from './lib/components/thank-you/lovedOneThankYou.component';
 export * from './lib/components/password/password.component';
 export * from './lib/components/password/password-redesigned.component';
 export * from './lib/components/redirect-to-login-landing/redirectToLoginLanding.component';
-export * from './lib/components/stay-informed/stayInformed.component';
+export * from './lib/components/stay-informed/stay-informed.component';
+export * from './lib/components/stay-informed/stay-informed-redesigned.component';
 export * from './lib/components/warning-message/warning-message.component';
 export * from './lib/components/activity/activity-link.component';
 export * from './lib/components/international-patients/internationalPatients.component';
